+++ src/main/java/nc/noumea/mairie/abs/service/impl/RecuperationService.java
@@ -79,7 +79,7 @@ public class RecuperationService implements IRecuperationService {
 	public SoldeDto getAgentSoldeRecuperation(Integer idAgent) {
 		AgentRecupCount soldeRecup = recuperationRepository.getAgentRecupCount(idAgent);
 		SoldeDto dto = new SoldeDto();
-		dto.setSolde(soldeRecup == null ? 0 : soldeRecup.getTotalMinutes());
+		dto.setSolde((double) (soldeRecup == null ? 0 : soldeRecup.getTotalMinutes()));
 		return dto;
 	}
